Split expression package into files by semantic concern
Pure code motion: every declaration moves verbatim with its doc comment,
verified by line-multiset comparison against the original file. No API,
naming, or behavior change; expression_test.go is untouched.

expression.go (251 lines) keeps the package doc, owned source types,
parsing front doors (Parse, parseCondition, expressionBody,
referencePath, visitTemplateExpressions), and helpers shared across
concerns (conditionNumber, decodeJSONValue, objectValue,
containsStatusFunction, isJobStatusReference).

New files:
- compile.go: CompileContext, graph-time substitution and evaluation
- condition.go: ConditionContext/ConditionScope, condition validation,
  strict fail-closed condition evaluation
- action_input_default.go: the loose-coercion action input default
  family, kept apart from the strict condition family by design
- runtime.go: direct runtime template evaluation and reference
  resolution
- references.go: the narrow reference-introspection predicates

Noted while splitting: validateRuntimeTemplate has no production
callers, only tests; left in place because this change is motion only.

Amp-Thread-ID: https://ampcode.com/threads/T-019ff0ab-0530-75ba-a946-8c2e9e9544ca
